fix integer and double inconsistencies

+++ include/machine_learning_algos/LinearRegression.h
@@ -4,9 +4,9 @@ class LinearRegression {
     public:
         LinearRegression(double learning_rate = 0.01, double threshold = 0.00001) : learning_rate(learning_rate), threshold(threshold) {};
 
-        void fit(const std::vector<std::vector<double>>& X, const std::vector<int>& y);
+        void fit(const std::vector<std::vector<double>>& X, const std::vector<double>& y);
 
-        std::vector<int> predict(const std::vector<std::vector<double>>& X);
+        std::vector<double> predict(const std::vector<std::vector<double>>& X);
     private:
         enum WeightType {
             BIAS,
@@ -22,7 +22,7 @@ class LinearRegression {
         double learning_rate;
         double threshold;
 
-        double MSE_loss(const std::vector<int>& y_true, const std::vector<int>& y_pred);
+        double MSE_loss(const std::vector<double>& y_true, const std::vector<double>& y_pred);
 
         
 };
